requirements.txt
alembic==1.8.1
autopep8==1.7.0
bcrypt==4.0.0
black==22.12.0
blinker==1.5
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.2.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
Flask-WTF==1.0.1
greenlet==1.1.3
gunicorn
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.0
pycodestyle==2.9.1
python-decouple==3.6
SQLAlchemy==1.4.40
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1